New easier on the users updates install script
This is three new script files.

computerNeedsReboot.sh - checks for the file is set by this or other
scripts to see if the computer needs a reboot. 1 is yes, 0 is no. It
will then display a messaging letting the user know they need to
reboot.

justInstallUpdates.sh - is the make working script. It will install the
updates on the computer using a policy. It will also set the file if
the machine needs a reboot or not.

needRebootReset.sh - is a script to run at system startup to reset the
needs reboot file.